python/2451i/journal/requirements.txt
mysql-connector-python
ppretty
prettytable
